+++ src/main/java/xyz/lunala/meowstorage/block/chests/DiamondChestBlock.java
@@ -4,7 +4,7 @@ import net.minecraft.world.item.ItemStack;
 import net.minecraft.world.level.block.entity.BlockEntityType;
 import net.minecraft.world.level.block.state.BlockState;
 import net.minecraft.world.level.storage.loot.LootParams;
-import xyz.lunala.meowstorage.block.entity.DiamondChestBlockEntity;
+import xyz.lunala.meowstorage.block.entity.chests.DiamondChestBlockEntity;
 import xyz.lunala.meowstorage.init.BlockEntityInit;
 import xyz.lunala.meowstorage.init.ItemInit;
