+++ src/main/java/xyz/lunala/meowstorage/block/entity/backpacks/BigBackpackEntity.java
@@ -1,7 +1,8 @@
-package xyz.lunala.meowstorage.block.entity;
+package xyz.lunala.meowstorage.block.entity.backpacks;
 
 import net.minecraft.core.BlockPos;
 import net.minecraft.world.level.block.state.BlockState;
+import xyz.lunala.meowstorage.block.entity.MeowBackpackEntityBase;
 import xyz.lunala.meowstorage.init.BlockEntityInit;
 
 public class BigBackpackEntity extends MeowBackpackEntityBase {
