+++ src/main/java/xyz/lunala/meowstorage/block/entity/barrels/CopperBarrelBlockEntity.java
@@ -1,4 +1,4 @@
-package xyz.lunala.meowstorage.block.entity;
+package xyz.lunala.meowstorage.block.entity.barrels;
 
 import net.minecraft.core.BlockPos;
 import net.minecraft.world.level.block.entity.BlockEntity;
