+++ src/main/java/xyz/lunala/meowstorage/block/entity/chests/CopperChestBlockEntity.java
@@ -1,22 +1,21 @@
-package xyz.lunala.meowstorage.block.entity;
+package xyz.lunala.meowstorage.block.entity.chests;
 
 import net.minecraft.core.BlockPos;
-import net.minecraft.network.chat.Component;
 import net.minecraft.world.level.block.state.BlockState;
-import xyz.lunala.meowstorage.init.BlockEntityInit;
 
-import static xyz.lunala.meowstorage.Meowstorage.MODID;
+import xyz.lunala.meowstorage.block.entity.MeowChestEntityBase;
+import xyz.lunala.meowstorage.init.BlockEntityInit;
 
 /**
  * Block entity for the Diamond Chest.
  * Extends MeowChestBaseBlockEntity to inherit common chest block entity functionalities.
  */
-public class DiamondChestBlockEntity extends MeowChestEntityBase {
+public class CopperChestBlockEntity extends MeowChestEntityBase {
 
     // The number of inventory slots for the Diamond Chest.
-    public static final int INVENTORY_SIZE = 432;
+    public static final int INVENTORY_SIZE = 54;
     // The display title for the Diamond Chest's GUI.
-    private static final String MATERIAL = "diamond";
+    private static final String MATERIAL = "copper";
 
     /**
      * Constructor for the DiamondChestBlockEntity.
@@ -25,7 +24,7 @@ public class DiamondChestBlockEntity extends MeowChestEntityBase {
      * @param pos The BlockPos of the block entity.
      * @param state The BlockState of the block entity.
      */
-    public DiamondChestBlockEntity(BlockPos pos, BlockState state) {
-        super(BlockEntityInit.DIAMOND_CHEST.get(), pos, state, INVENTORY_SIZE, MATERIAL);
+    public CopperChestBlockEntity(BlockPos pos, BlockState state) {
+        super(BlockEntityInit.COPPER_CHEST.get(), pos, state, INVENTORY_SIZE, MATERIAL);
     }
 }
