+++ src/main/java/xyz/lunala/meowstorage/block/entity/chests/DiamondChestBlockEntity.java
@@ -1,20 +1,20 @@
-package xyz.lunala.meowstorage.block.entity;
+package xyz.lunala.meowstorage.block.entity.chests;
 
 import net.minecraft.core.BlockPos;
 import net.minecraft.world.level.block.state.BlockState;
-
+import xyz.lunala.meowstorage.block.entity.MeowChestEntityBase;
 import xyz.lunala.meowstorage.init.BlockEntityInit;
 
 /**
  * Block entity for the Diamond Chest.
  * Extends MeowChestBaseBlockEntity to inherit common chest block entity functionalities.
  */
-public class CopperChestBlockEntity extends MeowChestEntityBase {
+public class DiamondChestBlockEntity extends MeowChestEntityBase {
 
     // The number of inventory slots for the Diamond Chest.
-    public static final int INVENTORY_SIZE = 54;
+    public static final int INVENTORY_SIZE = 432;
     // The display title for the Diamond Chest's GUI.
-    private static final String MATERIAL = "copper";
+    private static final String MATERIAL = "diamond";
 
     /**
      * Constructor for the DiamondChestBlockEntity.
@@ -23,7 +23,7 @@ public class CopperChestBlockEntity extends MeowChestEntityBase {
      * @param pos The BlockPos of the block entity.
      * @param state The BlockState of the block entity.
      */
-    public CopperChestBlockEntity(BlockPos pos, BlockState state) {
-        super(BlockEntityInit.COPPER_CHEST.get(), pos, state, INVENTORY_SIZE, MATERIAL);
+    public DiamondChestBlockEntity(BlockPos pos, BlockState state) {
+        super(BlockEntityInit.DIAMOND_CHEST.get(), pos, state, INVENTORY_SIZE, MATERIAL);
     }
 }
