+++ src/main/java/xyz/lunala/meowstorage/block/entity/chests/NetheriteChestBlockEntity.java
@@ -1,26 +1,9 @@
-package xyz.lunala.meowstorage.block.entity;
+package xyz.lunala.meowstorage.block.entity.chests;
 
 import net.minecraft.core.BlockPos;
-import net.minecraft.core.Direction;
-import net.minecraft.nbt.CompoundTag;
-import net.minecraft.network.chat.Component;
-import net.minecraft.world.MenuProvider;
-import net.minecraft.world.entity.player.Inventory;
-import net.minecraft.world.entity.player.Player;
-import net.minecraft.world.inventory.AbstractContainerMenu;
-import net.minecraft.world.level.block.entity.BlockEntity;
 import net.minecraft.world.level.block.state.BlockState;
-import net.minecraftforge.common.capabilities.Capability;
-import net.minecraftforge.common.capabilities.ForgeCapabilities;
-import net.minecraftforge.common.util.LazyOptional;
-import net.minecraftforge.items.ItemStackHandler;
-import org.jetbrains.annotations.NotNull;
-import org.jetbrains.annotations.Nullable;
+import xyz.lunala.meowstorage.block.entity.MeowChestEntityBase;
 import xyz.lunala.meowstorage.init.BlockEntityInit;
-import xyz.lunala.meowstorage.menu.ChestMenu;
-import xyz.lunala.meowstorage.util.IChestBlockMenuProvider;
-
-import static xyz.lunala.meowstorage.Meowstorage.MODID;
 
 public class NetheriteChestBlockEntity extends MeowChestEntityBase {
     // The number of inventory slots for the Diamond Chest.
